refactor: split db-store sync into records, frontmatter, and test modules
